Separated out commands on main side.
