+++ packages/app/main/src/commands/settingsCommands.ts
@@ -1,0 +1,54 @@
+//
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license.
+//
+// Microsoft Bot Framework: http://botframework.com
+//
+// Bot Framework Emulator Github:
+// https://github.com/Microsoft/BotFramwork-Emulator
+//
+// Copyright (c) Microsoft Corporation
+// All rights reserved.
+//
+// MIT License:
+// Permission is hereby granted, free of charge, to any person obtaining
+// a copy of this software and associated documentation files (the
+// "Software"), to deal in the Software without restriction, including
+// without limitation the rights to use, copy, modify, merge, publish,
+// distribute, sublicense, and/or sell copies of the Software, and to
+// permit persons to whom the Software is furnished to do so, subject to
+// the following conditions:
+//
+// The above copyright notice and this permission notice shall be
+// included in all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED ""AS IS"", WITHOUT WARRANTY OF ANY KIND,
+// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+// LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+// WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+//
+
+import { CommandRegistry } from './commandRegistry';
+import { dispatch, getSettings } from '../settings';
+import { FrameworkSettings } from '@bfemulator/app-shared';
+
+/** Registers settings commands */
+export const registerCommands = () => {
+  // ---------------------------------------------------------------------------
+  // Saves global app settings
+  CommandRegistry.registerCommand('app:settings:save', (settings: FrameworkSettings): any => {
+    dispatch({
+      type: 'Framework_Set',
+      state: settings
+    });
+  });
+
+  // ---------------------------------------------------------------------------
+  // Get and return app settings from store
+  CommandRegistry.registerCommand('app:settings:load', (...args: any[]): FrameworkSettings => {
+    return getSettings().framework;
+  });
+};
